+++ CoreTestament/tests/compile-fail/DiscardedHandles.cpp
@@ -1,0 +1,9 @@
+#include <Testament/Testament.hpp>
+
+void discardTest() {
+    Testament::makeTest("discarded test", [] {});
+}
+
+void discardSuite() {
+    Testament::makeSuite("discarded suite");
+}